automatically backed up by learn

+++ index.js
@@ -33,5 +33,5 @@ function makeInt(n){
 }
 
 function preserveDecimal(n){
-  return parseInt(n)
+  return parseFloat(n)
 }
